refactor: remove uccrestbuilder folder and move under commands
