+++ splunk_add_on_ucc_framework/commands/rest_builder/endpoint/datainput.py
@@ -15,37 +15,38 @@
 #
 
 
-from splunk_add_on_ucc_framework.uccrestbuilder.endpoint.base import (
+from splunk_add_on_ucc_framework.commands.rest_builder.endpoint.single_model import (
     RestEndpointBuilder,
     RestEntityBuilder,
 )
 
 
-class SingleModelEntityBuilder(RestEntityBuilder):
-    def __init__(self, name, fields, **kwargs):
+class DataInputEntityBuilder(RestEntityBuilder):
+    def __init__(self, name, fields, input_type, **kwargs):
         super().__init__(name, fields, **kwargs)
+        self._input_type = input_type
 
     @property
     def name_spec(self):
-        return "<name>"
+        return f"{self._input_type}://<name>"
 
     @property
     def name_default(self):
-        return "default"
+        return self._input_type
 
     @property
     def name_rh(self):
         return ""
 
 
-class SingleModelEndpointBuilder(RestEndpointBuilder):
+class DataInputEndpointBuilder(RestEndpointBuilder):
 
     _rh_template = """
 from splunktaucclib.rest_handler.endpoint import (
     field,
     validator,
     RestModel,
-    SingleModel,
+    DataInputModel,
 )
 from splunktaucclib.rest_handler import admin_external, util
 from {handler_module} import {handler_class}
@@ -55,10 +56,10 @@ util.remove_http_proxy_env_vars()
 
 {entity}
 
-endpoint = SingleModel(
-    '{conf_name}',
+
+endpoint = DataInputModel(
+    '{input_type}',
     model,
-    config_name='{config_name}'
 )
 
 
@@ -70,6 +71,14 @@ if __name__ == '__main__':
     )
 """
 
+    def __init__(self, name, namespace, input_type, **kwargs):
+        super().__init__(name, namespace, **kwargs)
+        self.input_type = input_type
+
+    @property
+    def conf_name(self):
+        return "inputs"
+
     def actions(self):
         return ["edit", "list", "remove", "create"]
 
@@ -79,6 +88,5 @@ if __name__ == '__main__':
             handler_module=self.rh_module,
             handler_class=self.rh_class,
             entity=entity.generate_rh(),
-            conf_name=self.conf_name,
-            config_name=self._name,
+            input_type=self.input_type,
         )
