+++ splunk_add_on_ucc_framework/rest_conf.py
@@ -15,7 +15,9 @@
 #
 from typing import Sequence
 
-from splunk_add_on_ucc_framework.uccrestbuilder.endpoint.base import RestEndpointBuilder
+from splunk_add_on_ucc_framework.commands.rest_builder.endpoint.base import (
+    RestEndpointBuilder,
+)
 
 
 class RestmapConf:
